backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
requests==2.31.0
pydantic[email]==2.5.0
pyjwt>=2.8.0,<3.0.0